Added cross-validation summary table. Introduced GBM cross-validation tests.
