+++ h2o-py/tests/testdir_generic_model/pyunit_generic_model_mojo_gbm.py
@@ -39,7 +39,7 @@ def compare_regression_output(original, generic):
 
 def test(x, y, output_test):
     airlines = h2o.import_file(path=pyunit_utils.locate("smalldata/testng/airlines_train.csv"))
-    gbm = H2OGradientBoostingEstimator(ntrees=1)
+    gbm = H2OGradientBoostingEstimator(ntrees=1, nfolds=3)
     gbm.train(x=x, y=y, training_frame=airlines, validation_frame=airlines)
     print(gbm)
     with Capturing() as original_output:
